Refactor: move single-use types to their respective files
Move types that are only imported by one file into that file and remove exports. This improves code locality and reduces unnecessary coupling.

Changes:
- Move GameOptions and GameController from src/types.ts to src/core/game.ts
- Move InputAdapter from src/adapters/types.ts to src/adapters/input-adapter.ts
- Move SoundAdapter from src/adapters/types.ts to src/adapters/sound-adapter.ts
- Move TerminalDimensions and TerminalAdapter from src/adapters/types.ts to src/adapters/terminal-adapter.ts
- Update demo adapter imports to use correct type locations
- Remove all "backwards compatibility" re-exports
