+++ src/adapters/types.ts
@@ -24,45 +24,3 @@ export interface Key {
  */
 export type KeypressHandler = (ch: string, key: Key) => void;
 
-/**
- * Input adapter interface for keyboard input
- */
-export interface InputAdapter {
-	initialize: () => void;
-	onKeypress: (handler: KeypressHandler) => void;
-	cleanup: () => void;
-}
-
-// ============================================================================
-// Sound Adapter Types
-// ============================================================================
-
-/**
- * Sound adapter interface for audio playback
- */
-export interface SoundAdapter {
-	play: (soundName: string) => void;
-	toggleMute: () => boolean;
-	isMuted: () => boolean;
-}
-
-// ============================================================================
-// Terminal Adapter Types
-// ============================================================================
-
-/**
- * Terminal dimensions
- */
-export interface TerminalDimensions {
-	width: number;
-	height: number;
-}
-
-/**
- * Terminal adapter interface for console output
- */
-export interface TerminalAdapter {
-	clear: () => void;
-	log: (message: string) => void;
-	getDimensions: () => TerminalDimensions;
-}
